fix: split mock preload into server and utils
