+++ packages/ipfs/test/core/preload.spec.js
@@ -5,7 +5,7 @@ const { nanoid } = require('nanoid')
 const { Buffer } = require('buffer')
 const { expect } = require('interface-ipfs-core/src/utils/mocha')
 const all = require('it-all')
-const MockPreloadNode = require('../utils/mock-preload-node')
+const MockPreloadNode = require('../utils/mock-preload-node-utils')
 const IPFS = require('../../src')
 const createTempRepo = require('../utils/create-repo-nodejs')
